Inserção dos comandos de contagem de tempo,trocas e  retorno por objeto da classe Performance nas classes BubbleSort e insertionSort. Alteração do tipo da variavel  ( de int para long) tempExecucao da classe Performace .

+++ src/Model/PerformanceDoMetodo.java
@@ -3,7 +3,7 @@ package Model;
 public class PerformanceDoMetodo {
 	
 	private int qtdTrocas;
-	private int tempeExecucao;
+	private long tempeExecucao;
 	private String nomeAlgoritmo;
 	
 	public PerformanceDoMetodo(){
@@ -20,11 +20,11 @@ public class PerformanceDoMetodo {
 		this.qtdTrocas = qtdTrocas;
 	}
 
-	public int getTempeExecucao() {
+	public long getTempeExecucao() {
 		return tempeExecucao;
 	}
 
-	public void setTempeExecucao(int tempeExecucao) {
+	public void setTempeExecucao(long tempeExecucao) {
 		this.tempeExecucao = tempeExecucao;
 	}
